+++ apps/mrpt_apps_gui/robot-map-gui/gui/configWidget/CGasGridConfig.cpp
@@ -34,23 +34,27 @@ void CGasGridConfig::updateConfiguration(mrpt::maps::TMetricMapInitializer* opti
   ASSERT_(mapDefination);
 
   Q_UNUSED(options);
-  mapDefination->min_x = m_ui->min_x->value();
-  mapDefination->max_x = m_ui->max_x->value();
-  mapDefination->min_y = m_ui->min_y->value();
-  mapDefination->max_y = m_ui->max_y->value();
-  mapDefination->resolution = m_ui->resolution->value();
+  mapDefination->min_x = static_cast<float>(m_ui->min_x->value());
+  mapDefination->max_x = static_cast<float>(m_ui->max_x->value());
+  mapDefination->min_y = static_cast<float>(m_ui->min_y->value());
+  mapDefination->max_y = static_cast<float>(m_ui->max_y->value());
+  mapDefination->resolution = static_cast<float>(m_ui->resolution->value());
 
   mapDefination->insertionOpts.gasSensorLabel = m_ui->gasSensorLabel->text().toStdString();
-  mapDefination->insertionOpts.enose_id = m_ui->enose_id->value();
-  mapDefination->insertionOpts.gasSensorType = m_ui->gasSensorType->value();
+  mapDefination->insertionOpts.enose_id = static_cast<uint16_t>(m_ui->enose_id->value());
+  mapDefination->insertionOpts.gasSensorType = static_cast<uint16_t>(m_ui->gasSensorType->value());
   mapDefination->insertionOpts.windSensorLabel = m_ui->windSensorLabel->text().toStdString();
 
   mapDefination->insertionOpts.useWindInformation = m_ui->useWindInformation->isChecked();
-  mapDefination->insertionOpts.advectionFreq = m_ui->advectionFreq->value();
-  mapDefination->insertionOpts.std_windNoise_phi = m_ui->std_windNoise_phi->value();
-  mapDefination->insertionOpts.std_windNoise_mod = m_ui->std_windNoise_mod->value();
-  mapDefination->insertionOpts.default_wind_direction = m_ui->default_wind_direction->value();
-  mapDefination->insertionOpts.default_wind_speed = m_ui->default_wind_speed->value();
+  mapDefination->insertionOpts.advectionFreq = static_cast<float>(m_ui->advectionFreq->value());
+  mapDefination->insertionOpts.std_windNoise_phi =
+      static_cast<float>(m_ui->std_windNoise_phi->value());
+  mapDefination->insertionOpts.std_windNoise_mod =
+      static_cast<float>(m_ui->std_windNoise_mod->value());
+  mapDefination->insertionOpts.default_wind_direction =
+      static_cast<float>(m_ui->default_wind_direction->value());
+  mapDefination->insertionOpts.default_wind_speed =
+      static_cast<float>(m_ui->default_wind_speed->value());
 }
 
 TypeOfConfig CGasGridConfig::type() const { return TypeOfConfig::GasGrid; }
